+++ CHANGELOG.md
@@ -452,3 +452,8 @@
 - there may be a *different* off by one error. '*During* cycle N...'
 - to test that theory, I output regX's History cycles 19..21, etc
 - it *does* look like I am off by one!
+
+---
+
+- correct answer from the test data
+- off by one! lookup cycle 19, 59, 99 ...